Fixed the naming conventions of the old site so that images & links work
IMPORTANT!: Make sure that all of your naming conventions are consistent. For this commit, I had to change the upper-cased banner files to lowercase b/c that's how they were referenced in the html. same w/ the button in the unit four of the home page. It might be a by/browser/basis, so chrome was commentating by not caring, but some browsers do care so make sure you're spelling things correct.